requirements.txt
pyside6
bdfparser
pillow
numpy